pyproject.toml
[project]
name = "cronkite"
version = "0.1.0"
description = "LLM-powered news aggregation engine"
authors = [
    {name = "Euan", email = "euan.m.campbell@protonmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[tool.poetry]
packages = [{include = "cronkite", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
